+++ lib/core/constants/legal_urls.dart
@@ -1,0 +1,20 @@
+/// 법적 문서 URL 상수
+///
+/// 앱스토어 제출 전 실제 URL로 변경 필요
+abstract class LegalUrls {
+  /// 개인정보 처리방침 URL
+  /// TODO: 실제 호스팅 URL로 변경
+  static const String privacyPolicy = 'https://YOUR_DOMAIN.com/privacy-policy';
+
+  /// 이용약관 URL
+  /// TODO: 실제 호스팅 URL로 변경
+  static const String termsOfService = 'https://YOUR_DOMAIN.com/terms-of-service';
+
+  /// 건강정보 면책조항 URL
+  /// TODO: 실제 호스팅 URL로 변경
+  static const String medicalDisclaimer = 'https://YOUR_DOMAIN.com/medical-disclaimer';
+
+  /// 고객 지원 이메일
+  /// TODO: 실제 이메일로 변경
+  static const String supportEmail = 'support@YOUR_DOMAIN.com';
+}